Move saved search file to util folder
